requirements.txt
# python -m venv .venv

# pip install -r .\requirements.txt

fastapi
psycopg2
python-dotenv
uvicorn
asyncio
asyncpg
sqlalchemy


# uvicorn main:app --reload

# uvicorn main:app --reload --host localhost